backend/requirements.txt
fastapi
uvicorn
python-multipart
whisper
ffmpeg-python
yt-dlp
google-generativeai
pydantic
